Fix plug-in start on Mac -> moved Characteristics back to accessory.js
