+++ base/accessory.js
@@ -2,12 +2,6 @@ const ping = require('ping');
 const moment = require('moment');
 const arp = require('node-arp');
 
-const {
-  LastActivationCharacteristic,
-  SensitivityCharacteristic,
-  DurationCharacteristic,
-} = require('./characteristics');
-
 class PeopleProAccessory {
   constructor(log, config, platform) {
     this.log = log;
@@ -34,6 +28,59 @@ class PeopleProAccessory {
       this.service
         .getCharacteristic(Characteristic.MotionDetected)
         .on('get', this.getState.bind(this));
+
+      class LastActivationCharacteristic extends Characteristic {
+        constructor() {
+          super('LastActivation', 'E863F11A-079E-48FF-8F27-9C2605A29F52');
+          this.setProps({
+            format: Characteristic.Formats.UINT32,
+            unit: Characteristic.Units.SECONDS,
+            perms: [
+              Characteristic.Perms.READ,
+              Characteristic.Perms.NOTIFY,
+            ],
+          });
+        }
+      }
+
+      class DurationCharacteristic extends Characteristic {
+        constructor() {
+          super('Duration', 'E863F12D-079E-48FF-8F27-9C2605A29F52');
+          this.setProps({
+            format: Characteristic.Formats.UINT16,
+            unit: Characteristic.Units.SECONDS,
+            minValue: 5,
+            maxValue: 15 * 3600,
+            validValues: [
+              5, 10, 20, 30,
+              1 * 60, 2 * 60, 3 * 60, 5 * 60, 10 * 60, 20 * 60, 30 * 60,
+              1 * 3600, 2 * 3600, 3 * 3600, 5 * 3600, 10 * 3600, 12 * 3600, 15 * 3600,
+            ],
+            perms: [
+              Characteristic.Perms.READ,
+              Characteristic.Perms.NOTIFY,
+              Characteristic.Perms.WRITE,
+            ],
+          });
+        }
+      }
+
+      class SensitivityCharacteristic extends Characteristic {
+        constructor() {
+          super('Sensitivity', 'E863F120-079E-48FF-8F27-9C2605A29F52');
+          this.setProps({
+            format: Characteristic.Formats.UINT8,
+            minValue: 0,
+            maxValue: 7,
+            validValues: [0, 4, 7],
+            perms: [
+              Characteristic.Perms.READ,
+              Characteristic.Perms.NOTIFY,
+              Characteristic.Perms.WRITE,
+            ],
+          });
+        }
+      }
 
       this.service.addCharacteristic(LastActivationCharacteristic);
       this.service
