--- base/characteristics/index.js
@@ -1,9 +1,0 @@
-const LastActivationCharacteristic = require('./last_activation');
-const SensitivityCharacteristic = require('./sensitivity');
-const DurationCharacteristic = require('./duration');
-
-module.exports = {
-  LastActivationCharacteristic,
-  SensitivityCharacteristic,
-  DurationCharacteristic,
-};
