+++ nodemon.json
@@ -1,0 +1,13 @@
+{
+  "watch": [
+    "base",
+    "index.js"
+  ],
+  "ext": "js",
+  "ignore": [],
+  "exec": "homebridge -I -D",
+  "signal": "SIGTERM",
+  "env": {
+    "NODE_OPTIONS": "--trace-warnings"
+  }
+}
